[Extensions c2s][Reland] Add ExtensionsTabbedMenuView skeleton
Original CL description:
Introduces the extensions tabbed menu structure (dialog with a tabbed
pane), without any content. For now, this menu is only opened by
ExtensionsToolbarButton under a flag.

Reverted reason:
Interactive UI test was flaky because the button click event was not
always in the correct point. This is due to AnimatingLayout
causing weird behaviors in the container layout that still need to be
fixed.

Fix:
Changed interactive ui test to unit test, which in general should be
preferred since they are faster and less flaky.

Bug: 1263311
Change-Id: I45df0b56a3b0ba4dee2ff670549d92afec8bc982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3294045
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944217}