[Reland] Implement Calendar events expanded view
The original cl was reverted due to this crash report:
https://ci.chromium.org/ui/p/chromium/builders/ci/Linux%20ChromiumOS%20MSan%20Tests/25454/test-results

Fix in this cl:
It seems that when the `CalendarView` is destructed, it first reset the
controller, then reset the `CalendarDateCellView`. In this cl, the
month view are removed from the CalendarView in its destructor, so that
these child views will be removed from the controller first. See
the changed file: calendar_view.cc


----------Previous description---------
In this cl the implementation includes:
* click on the date cell will open the expanded view
  - draw a solid background for the selected date cell
  - shows the selected date's event list.
  - Auto scroll to the selected date's row
  - disable the up/down/today button function
  - disable the main calendar view scrollable function
* click on another date cell when the expanded view is expanded already
  - updates the event list
  - remove the solid background of the previous date cell
* click on the close button will close the expanded view

Several TODOs need to be done in the next cl:
* CalendarEventListItemView polish
  - colored dot
  - time label
  - Paddings
* animation
* enable calendar view up/down button to scroll one week when the event
list is expanded.

See screen record:
https://drive.google.com/file/d/1yBt63-Z-WYYFPBJ3VRM1p8NkN0mpeXvq/view?usp=sharing
which are the corresponding meetings here:
https://screenshot.googleplex.com/8bwAouUaTWHNgcD in the google calendar,
we can see the start and end time is converted to the system local time
in the chrome calendar view.

Bug: 1238927
Change-Id: I5c918665d099366955ed92ea1d7f438ae11eca71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296666
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944328}