What's New: Don't open a new tab if there is no content
Instead of opening and then closing the tab if the network request
fails, only open the tab after confirming the page exists and can be
reached.

Update the URL to load to reflect the latest server side logic, which
includes a way to request a page that redirects to the closest page
to the current version.

Since we no longer need to worry about page load failures due to the
server side changes, remove the error page and associated prefetch
logic that was handling this previously.

Also, instead of showing What's New as a background tab (which we don't
ever want to do), avoid showing it for cases where there is a reset
profile or welcome back tab.

Bug: 1267819
Change-Id: Ibd31506d6a772cc6d00cf959c00a5193a1734157
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3275312
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944332}