Fixed browser from crashing when permanent bookmark nodes are clicked when the tab group save flag is enabled.
If our bookmark node is a permanent one (other, overflow/hidden, mobile),
we continue to use the current implementation which opens a context
menu on left click. Otherwise, all other bookmark nodes will be updated
with the tab groups save functionality (opening all tabs on click in a
new tab group).

It should be noted that in an attached crash report for this bug,
https://listnr.corp.google.com/product/208/report/89091548209,
there is mention of crashing when clicking a non permanent node.
This has been addressed in the following change
https://chromium-review.googlesource.com/c/chromium/src/+/3282779.

Bug: 1271684
Change-Id: I3c1588462efd46ae36e9c823f9be59e88ef0eced
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297132
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944378}