[Files SWA] Handle destination folder in IsCrossFileSystem
Follow-up to crrev.com/c/3288609: IsCrossFileSystem should now handle
the destination folders for copy/move operations, not only the paths for
each individual destination entries.

In particular, the Downloads folder is a valid destination, and should
be handled as "On the Downloads volume and in the Downloads
bind-mounted folder".

Bug: b:204838286
Change-Id: I533f065ab514a88e29af87336873773e1810067d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3292913
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944380}