[Lacros] Fix extension signin with a stub account
ChromeOS uses a stub account when running in Linux VM for testing
without a login manager. IdentityManager doesn't return valid extended
info for this account which breaks an assumption in
ShowExtensionSigninPrompt() that the user is always signed-in on Lacros.

This issue can be fixed by not asking for an extended account info in
IdentityGetAuthTokenFunction::ShowExtensionLoginPrompt().
User's email is already available in ShowExtensionLoginPrompt(), so
there is no need in an extra fetch.

Another solution would be to return early when the user is signed-in
with the stub account. However, this requires a UserManager object that
is not available on Lacros.

Fixed: 1253768
Change-Id: I21c371fd075cc75b35aa152c389ebb8de8410f29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295469
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944445}