[FRE] Make the new signin FRE page scrollable
This cl adjusts the layout of the new signin FRE page to be scrollable
when the screen size is too small or if chrome is launched in split
screen mode.

In Addition this cl adds logic to only show the landscape mode UI if the
device width is at least 600dp since the landscape UI is a two pane
view.

Bug: 1261734
Change-Id: Ie065a85ff820d218f6ffce896dffd4269b0a918e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3284807
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944510}