[Profile picker] Enable browser tests on lacros
This CL enabled the first batch of browser tests for profile picker and
for profile creation flow that make sense on lacros. Follow-up CLs will
enable more tests and add lacros-specific tests.

This CL makes a change in the turn sync on helper (disabling all data
types) that is needed for the tests that choose not to sync to detect
that everything works as intended.

Bug: 1248047
Change-Id: I6d618800f499502ce8be80e17a0177366865aeeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3234519
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944518}