[send-tab-to-self desktop] Delete ShouldOfferFeatureForLink()
Before this CL, ShouldOfferFeatureForLink() checked the scheme of the
current URL besides that of the link to be shared. The reason is unknown
and this is possibly a bug, since the current URL never makes it to the
other client anyway. This CL deletes ShouldOfferFeatureForLink() and
starts calling ShouldOfferFeatureForPage() with the link URL instead.
The latter is renamed to ShouldOfferToSendUrl(), to better reflect its
new use.

This change is actually a no-op if the share::kShareMenu feature toggle
is enabled, because that removes STTS for links, c.f. crbug.com/1273150.
The true usefulness of this change is reducing the number of predicates
of type "Is STTS allowed", so we can later replace them with a single
method that tells in which of the 3 states the STTS bubble is:
a) feature enabled, choose target device
b) feature enabled but no target device
c) feature disabled, offer sign-in promo

We also use this opportunity for some code improvements:
- Drop the IsIncognitoProfile() check in ShouldOfferFeatureForPage().
This was redundant with the null check of SendTabToSelfSyncService.
Worse, it didn't cover the Guest Profile case, where the service is
also null. Now the method doesn't depend on profile and can be tested
more easily.
- Delete IsUserSyncTypeActive() and HasValidTargetDevice(), which
were exposed in send_tab_to_self_util.h, but were never directly
called in production code, only in tests.

Bug: 1219433
Change-Id: Iba5cfa4a2cf47942fd220e8660e7a57e97a0af14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295397
Commit-Queue: Victor Vianna <victorvianna@google.com>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944610}