Remove SortButton that was used for sort testing
The SortButton in AppsContainerView was a button class that shows arrow
buttons in tablet mode that can sort the app list on click/tap. As the
context menu for sorting is implemented, remove this class and replace
the functionality by using the context menu on AppsGridView and
AppListItemView.

Bug: 1263999
Change-Id: I8de26ab8186af4332195063617912529db15e46d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3290605
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944623}