[ios] Refine when NTP offset is saved in Single NTP feature
By setting the NTPCoordinator's WebState to nullptr when the NTP
visibility goes to inactive, then the offset saved for page navs
matches what happens in the default behavior.

Adding a NTP URL check for the single NTP feature in
saveContentOffsetForWebState: will also fix a bug where the offset
was being saved in instances where the NTP was not the current page.

Adding an equality check in NTPCoordinator reduces the redundancy
in which the offset saving logic gets called.

Bug: 1272060
Change-Id: Ide6cae7fa5b89bb9c7a2acf5112eb77bae6a24df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3293920
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#944653}