Respect explicit flat pixel structure.
The PixelStructure registry entry defines the value 0 to mean the
display should not have ClearType applied (flat). This allows for
ClearType to be enabled in general but disabled explicitly for a
display. This value should be respected if possible.

In the event an unknown value is used by the PixelStructure registry
entry, this should be treated as flat.

In the event that ClearType is enabled but there is no PixelStructure
registry entry, fall back to the SPI font smoothing orientation. If
even this value can not be read, default to flat.

Bug: chromium:434780
Change-Id: Ibde7111f463c02036a969beb81e2aa51e3c48ab9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296543
Commit-Queue: Ben Wagner <bungeman@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944668}