[ZT] Expose DTKeyManager Is Fully Initialized in Internals Page
On the chrome://connectors-internals page, a new row will let the user
know if DTKeyManager has successfully loaded a key and is ready for
handling attestation flows. If it's not, it will show false. On
unsupported platforms (e.g. Ash), the row will be hidden.

Bug: b/207511801
Change-Id: Id745065f81e6c595303bfbe20443009256a5b726
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297999
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944683}