Redefine the system token software fallback
Only doing the software fallback on system token would cause chromium
to crash when it's trying to initialize the user token. Redefine the
system slot software fallback to NSS slots software fallback could
resolve this issue.
Because the hardware NSS slots would only be available when the TPM is
owned, we should use the fallback when the TPM isn't owned.

BUG=b:204320629
TEST=Build with nss_slots_software_fallback = true
TEST=disallow the TPM on the device, and check chrome://flags

Change-Id: Iac7ad4bdeb5905de3f1022943a652636983bff9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3173937
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Commit-Queue: Yi Chou <yich@google.com>
Cr-Commit-Position: refs/heads/main@{#944692}