ash: Fix launcher continue section layout after search updates
Updates to the search results would cause an unlimited number of rows
(and padding rows) to be added to the TableLayout. This created an
expanding empty space in the middle of the launcher.

Since the continue section only shows if there are 3 or 4 tasks, the
number of rows is fixed. Add the rows to the TableLayout once, at
initialization time.

If we need to change this behavior we'll need to reset the TableLayout
layout manager on Update(), since it does not have an API to remove
rows.

Bug: 1273170
Test: added to ash_unittests
Change-Id: Ibbf5423cc42baf152cb87964806c1b8d79412d44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3299140
Reviewed-by: Ana Salazar <anasalazar@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944740}