content: updates test to wait for DidStopLoading()
This test creates a subframe in such a way that it navigates to
about:blank and at the same time initiates a navigation in the
subframe. This patch makes the test wait for DidStopLoading (which is
generated from the renderer as part of the about:blank navigation)
before continuing. This way, when the test starts the navigation in
the frame there is no DidStopLoading() in flight that may impact the
test.

The beforeunload handler IPC effectively ensures DidStopLoading()
is received before starting the navigation. Removing the IPC means
the test needs to explicitly wait.

BUG=365039

Change-Id: I51bee4bd6ca8ffdbeec50ff1faae04294c0e918c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296330
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944768}