Fixed empty bookmark bar crashing when tab group save flag is enabled
The browser would crash when the tab group save flag was enabled and an empty bookmark was clicked. There was a DCHECK causing the browser to crash. Removing DCHECK entirely would prevent crashing but for safety we still check if the bookmark has items and terminate early in case there are extra operations that are performed inside of the OpenAllIfAllowed call.

Additionally, any links nested within a bookmark folder will not be opened.

Bug: 1271130
Change-Id: Idbd91832c09f3caf1e79b5c77c747e172e1ebb90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3299328
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944825}