Lacros: Enable lacros-arm-generic-rel on branches
This builder was not enabled on branches due to an oversight and this
CL fixes the problem.

Based on the capacity graph:
http://shortn/_picIofnUY5

There is abundant capacity to absorb the ~5% load increase.

This is a compile only bot, so has no risk of regressing cycle time.

Change-Id: I54ed6b81d5ad42a01074995dc8127ffb34436d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297368
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944850}