MPArch: Fenced frame should not reset a new tab start time in SearchTabHelper
SearchTabHelper::DidFinishLoad has been calculating the loading time of
an new tab page when the given render frame host is a main frame.

In terms of fenced frame, fenced frame should not affect the
loading time calculation because the calculation should be done for
a primary main frame. So, this CL replaces GetParent with
IsInPrimaryMainFrame in SearchTabHelper::DidFinishLoad, and add
a test case to check it.

Bug: 1271815
Change-Id: If9b2c064c7df9820c2cc339b45b8a531f8c566cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3291252
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Commit-Queue: Gyuyoung Kim <gyuyoung@igalia.com>
Cr-Commit-Position: refs/heads/main@{#944855}