🤠 Fix MimeUtilTest.ExtensionTest
`GetMimeTypeFromExtension` depends on some hardcoded list and on the
system where Chrome is running on. As a result, it fails on
android-arm-test and android-arm64-tests, where it gets overridden by
the system.

The error is:
```
../../net/base/mime_util_unittest.cc:62: Failure
Expected equality of these values:
  test.mime_type
    Which is: "application/x-mpegurl"
  mime_type
    Which is: "audio/x-mpegurl"
```
for the ".m3u8" extension.

What to do? Remove this case? Expect one of both?

Bug: 1273061
Change-Id: I410ce342e45102e7a54d35c9d7e732ff5a1db675
Fixed: 1273061
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3298129
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944895}