[StartupTabPreloader] Track reason for preload decision
StartupTabPreloader performs a preload at startup only 20-30% of the
time. This CL adds a histogram breaking down the reason for its decision
on whether to perform a preload or not.

As part of this change we slightly reorder the checks in
StartupTabPreloader#shouldLoadTab(): we move the check for whether the
intent specifies opening an incognito tab to before the check for
whether IntentHandler thinks that the intent should be ignored. The
reason is that IntentHandler specifies such intents should be ignored
when coming from non-Chrome apps, so having the more specific check
first allows us to capture this case separately in metrics. This change
will have no impact on production behavior, as the call to
IntentUtils.safeGetBooleanExtra() has no side-effects.

Bug: 1258621
Change-Id: I4fabdc1bcff13db9cf5f35a7782dbbfba50d76f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3296291
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Hao Dong <spdonghao@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944899}