Add DLP reporting for warning mode for printing
WARN report events are sent each time a warn dialog is displayed.

WARN_PROCEED events are sent each time a warn action is executed. Either
when a user selects to proceed with an action in the dialog or when the
action has been already accepted previously and the dialog is not shown.

Bug: 1227700
Change-Id: Icc43828805556e432947732eb6be1f040fba7f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3236733
Commit-Queue: Jan Kopański <jkopanski@google.com>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944937}