Not always dispatch touch event to content in AnchoredPopupWindow
The behavior introduced in [1] led to an issue with Autofill Dropdown
items not being clickable in M96 Stable on a subset of devices (see
video in [2]). It was needed to make the IPH popup content clickable
even if mDismissOnTouchInteraction is true.

In case of Autofill, mDismissOnTouchInteraction is false, but
dispatching the touch event to the content from the window's touch
handler sometimes breaks the event handling. Previously, the touch
event was handled by the dropdown content in a natural way.

The immediate fix is the following: dispatch the event only if
mDismissOnTouchInteraction is true. This will differentiate the Autofill
and IPH cases, and in case mDismissOnTouchInteraction is false the
content already handles the touch event naturally. However, we may want
to revisit this behavior for the AnchoredPopupWindow in general.

[1]https://chromium-review.googlesource.com/c/chromium/src/+/3159033/13/ui/android/java/src/org/chromium/ui/widget/AnchoredPopupWindow.java
[2]https://bugs.chromium.org/p/chromium/issues/detail?id=1267873

Bug: 1271646
Change-Id: I5917c65759e20b03b4dd97fe62279dc0fbf97471
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295575
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Matt Jones <mdjones@chromium.org>
Commit-Queue: Elizabeth Popova <lizapopova@google.com>
Cr-Commit-Position: refs/heads/main@{#944950}