[Lacros] Pick profile color for secondary accounts screen
This CL special cases color generation for the case where Chrome shows
the picker to select an account available as secondary for the current
profile. This flow is started from OGB in the current profile.

Instead of generating a random new color, the current color of the
profile is returned.

Bug: 1226076
Change-Id: I9057413896127333a6ebaf16d479e44b866a872f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3299229
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944977}