Fix logic for kEnableAutomation infobar
To determine whether or not the infobar that warns about the
kEnableAutomation flag is shown, the startup code used the command line
of the most recent invocation of Chromium. This was inconsistent
behavior, since the other code in Chromium only ever checks for presence
of this flag on the command line of the main browser process via
CommandLine::ForCurrentProcess().

Bug: 1271088
Change-Id: Ia4b09880ebb1f08641141351a786df0426ec1cb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3268668
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944981}