blink/platform: Avoid redirection in WebPrivatePtr
For garbage-collected types, WebPrivatePtr keeps a pointer to
a malloced-allocated (CrossThread)Persistent. This allocation (and
indirection) is unnecessary for Persistents, since Persistent can simply
be inlined in PtrStorageType. This should be beneficial for both memory
and performance.

As a consequence, since Persistents are two words in size (or three on
Debug builds), the CL introduces aligned_storage<sizeof(Persistent)>
into WebPrivatePtr, the second word of which is unused for
ref-counted types. We expect GCed WebPrivatePtrs to dominate, so the
overhead of storing an extra pointer for ref-counted WebPrivatePtrs
should be negligible.

Theoretically, we could dispatch between ref-counted and
garbage-collected types. However, on the public side some of the types
that WebPrivatePtr wraps are incomplete. The IsGarbageCollected<> trait
doesn't work for incomplete types.

The memory profile of Speedometer2 showed a huge number of allocations
that happened for this indirection layer (the top second location), half
of them are transient. This introduces a lot of pressure on PCScan. The
CL aims to improve it.

Current number of malloc-allocations coming from WebPrivatePtr:
Speedometer2: overall: 714684, transient: 384782
system_health.common_desktop, cnn:2021: overall: 95341, transient: 50153

Bug: 1249550
Change-Id: I4538e7978303825b3232c84b4f704821b781829a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3295949
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#944997}