[ZT] Add UMA Metrics for Attestation Funnel, Result and Latency
The Funnel histogram will not contain any form of error results, it only
accounts for how far the attestation flow has gotten.

The Result histogram will capture the final result of the attestation
step (validating challenge and returning a response) of the flow. On
Ash platforms, there is a simple mapping of Ash result code to the enum
for this new histogram.

The Latency histogram encompasses the latency for the attestation step.
It does not account for the initial back-and-forth HTTP requests, as
timing those would not be as useful (latency varies greatly between
devices).

Bug: b/206654279
Change-Id: I8321b6361e5021563db0a14f4cf40a1eceffc35d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3298169
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945021}