[PartitionAlloc][mac] Undo the initialization race fix.
PartitionAlloc's initialization is racy. Unfortunately, the fix in
https://chromium-review.googlesource.com/c/chromium/src/+/3141253
doesn't work, as it is not possible to copy a zone (which is not
documented, but a lot of this relies on undocumented hacks). The fix
might even widen the race window a bit more.

Remove the fix, and note in a comment that the race exists.

Bug: 1232392, 1272149
Change-Id: I0da15264668e94cfe0c2e8dfe8ed514a06ccb421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3297867
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945033}