Check kKioskMode on process command line instead of startup command line
While the old behavior did not result in bugs (as far as I can tell), it
is semantically more sound to read the flag from the command line of the
current process, because all other locations where the flag is used also
read it from the command line of the current process. This also aligns
its usage with `StartupBrowserCreatorImpl::MaybeToggleFullscreen`.

Change-Id: I552c0074c8199a950b9fa86506ddbdb76ab1860b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3298022
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#945039}